+++ src/colorPicker/palettesList/palettesList.tsx
@@ -1,10 +1,10 @@
 import React, { useCallback } from "react";
-import { PaletteObj } from "../types";
+import { StatePalette } from "../types";
 import Dropdown from "../dropdown/dropdown";
 import Palette from "./palette";
 
 type Props = {
-    palettes: PaletteObj[],
+    palettes: StatePalette[],
     current: number,
     onChange: (newCurrent: number) => void;
 };
@@ -21,10 +21,15 @@ const PalettesList = (props: Props) => {
         ? props.palettes[props.current].name
         : props.palettes[0].name
 
+    const itemProps = {
+        current: props.current,
+    }
+
     return (
         <Dropdown
             label={label || `Palette No${props.current + 1}`}
             items={props.palettes}
+            itemProps={itemProps}
             itemComponent={Palette}
             closeOnItemClick={true}
             renderList=">1"