+++ src/colorPicker/types.ts
@@ -1,27 +1,60 @@
+import { State } from '@storybook/api'
+
 export type ShadeType = {
-    label: string,
-    value: string,
+	label: string
+	value: string
 }
+
+export type ShadesType = Record<string, string> | string
+
+export type ColorPaletteAsObject = Record<string, ShadesType>
 
 export type ColorPaletteAsArray = {
-    label: string,
-    values: ShadeType[]
+	label: string
+	values: ShadeType[]
 }
 
-export type ShadesType = Record<string, string> | string;
-export type ColorPalette = Record<string, ShadesType> | ColorPaletteAsArray[];
-
-export type PaletteObj = {
-    name: string,
-    palette: ColorPalette,
+export type PaletteAsObject = {
+	name: string
+	palette: ColorPaletteAsObject
 }
+
+export type PaletteAsArray = {
+	name: string
+	palette: ColorPaletteAsArray[]
+}
+
+export type PaletteObj = PaletteAsObject | PaletteAsArray
 
 export type ColorPalettes = {
-    default: string,
-    palettes: PaletteObj[]
+	default: string
+	palettes: PaletteObj[]
 }
 
-export type AddonState = {
-    currentPalette: number,
-    applyColorTo?: string[],
-} | undefined
+export interface StatePalette {
+	name: string
+	palette: ColorPaletteAsArray[]
+}
+
+export interface StatePalettes {
+	default: string
+	palettes: StatePalette[]
+}
+
+export type StoryState = {
+	currentPalette: number
+	controls?: string[]
+	selectedControls: string[]
+	copyOnClick: boolean
+	storyPalettes: StatePalettes
+}
+
+export type StoryStates = Record<string, StoryState>
+
+export interface StorybookState extends State {
+	colorPalettes: StatePalettes
+}
+
+export interface AddonState {
+	storyStates: StoryStates
+}