+++ src/utils/index.tsx
@@ -1,0 +1,3 @@
+export * from "./getColorControls"
+export * from "./findDefayltPaletteIndex"
+export { default as transformPalette } from "./transformPalette"